% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFpower-package.R
\docType{package}
\name{BFpower-package}
\alias{BFpower}
\alias{BFpower-package}
\title{BFpower: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
