% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFpower-ttest.R
\name{BFpower.one.sample.t}
\alias{BFpower.one.sample.t}
\title{Calculate power for a one-sample Bayesian t-test with informative hypotheses}
\usage{
BFpower.one.sample.t(
  n = NULL,
  d = 0.1,
  sd = 1,
  threshold = 3,
  fraction = 1,
  power = NULL,
  hypothesis = "d=0",
  BFtype = c("AAFBF", "AFBF")
)
}
\arguments{
\item{n}{The sample size.}

\item{d}{The effect size.}

\item{sd}{The standard deviation.}

\item{threshold}{The threshold value for the Bayes factor (i.e., which Bayes
factor value is considered as convincing evidence for or against the hypothesis
of interest). By default, Bayes factors smaller than 1 are considered as evidence
against the hypothesis of interest.}

\item{fraction}{The fraction of information in the data used to construct the prior
distribution. The default value 1 denotes the minimal fraction, 2 denotes twice the
minimal fraction, and so on. See [bain::bain()] for more details.}

\item{power}{The target power of the test. If NULL, the function calculates the
power for the given sample size, effect size, and standard deviation.}

\item{hypothesis}{The hypothesis of interest. The default hypothesis denotes
that the effect size is equal to 0. See [bain::bain()] for more details.}

\item{BFtype}{The type of Bayes factor to calculate. Currently only the
approximate adjusted fractional Bayes factor is supported.}
}
\value{
The power of the test, or the sample size required to achieve said power.
}
\description{
Calculate power for a one-sample Bayesian t-test with informative hypotheses
}
